// src/types/films.ts
export interface Film {
    id: number
    title: string
    overview: string
    poster_path: string | null
    backdrop_path: string | null
    release_date: string
    vote_average: number
    vote_count: number
    genre_ids: number[]
    adult: boolean
    original_language: string
    original_title: string
    popularity: number
    video: boolean
  }
  
  export interface FilmDetails extends Film {
    genres: Genre[]
    runtime: number
    tagline: string
    budget: number
    revenue: number
    homepage: string
    imdb_id: string
    status: string
    production_companies: ProductionCompany[]
    production_countries: ProductionCountry[]
    spoken_languages: SpokenLanguage[]
  }
  
  export interface Genre {
    id: number
    name: string
  }
  
  export interface ProductionCompany {
    id: number
    logo_path: string | null
    name: string
    origin_country: string
  }
  
  export interface ProductionCountry {
    iso_3166_1: string
    name: string
  }
  
  export interface SpokenLanguage {
    english_name: string
    iso_639_1: string
    name: string
  }
  
  export interface TMDBResponse<T> {
    page: number
    results: T[]
    total_pages: number
    total_results: number
  }
  
  export type FilmCategory = 'popular' | 'top_rated' | 'upcoming'
  
  export interface WishlistItem {
    film: Film
    category: FilmCategory
    addedAt: string
  }